+++ signalflow/client.go
@@ -277,9 +277,16 @@ func (c *Client) keepReadingMessages(authenticatedCond *sync.Cond) {
 func (c *Client) acceptMessage(message messages.Message, authenticatedCond *sync.Cond) {
 	if _, ok := message.(*messages.AuthenticatedMessage); ok {
 		authenticatedCond.Signal()
-	} else {
-		log.Printf("Unknown SignalFlow message received: %v", message)
+		return
+	} else if msg, ok := message.(*messages.BaseJSONMessage); ok {
+		data := msg.RawData()
+		if data != nil && data["event"] == "KEEP_ALIVE" {
+			// Ignore keep alive messages
+			return
+		}
 	}
+
+	log.Printf("Unknown SignalFlow message received: %v", message)
 }
 
 func connect(ctx context.Context, streamURL *url.URL) (*websocket.Conn, error) {